Connection to Arduino
A launch file with a rosserial node has been added.
A arduino code made for an Arduino Mega. Has been made has weel.
A file python file converting the value of the controller to arduino should be added to perfect the connection between the two entities.
Servos for the gripper are not managed.